requirements.txt
# Core ML tooling
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2

# Crawling & parsing
requests==2.31.0
beautifulsoup4==4.12.2

# Embeddings & vector search
huggingface_hub==0.14.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# API & UI
fastapi==0.95.1
uvicorn[standard]==0.22.0
streamlit==1.25.0
